+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- Add exception tracing (#2)
+
 ## 0.2.7 (2020-10-23)
 
 - Adding `:protocol` option to connect config map